Added a few custom items and some prototypes for implementation.
The shutdown threading issue has been fixed.

Currently though, the ExplosiveWindCharge does not work. The PlayerInteractEvent is not handled at all, and I may need to change the KItem.handle architecture.

Fixed a Config bug where I was trying to add a value to a HashMap.KeySet (which is an unsupported operation. Not sure how I didn't catch that.)

Also changed custom items to Cookie to hopefully stop them from being used as crafting ingredients, as that is a hard system to work around. May look into ways to make custom items invalid in all recipes.

Took 2 hours 49 minutes